3.1.2rc8 (#10)
* send health every 10 pings instead of with every ping
* add additional checks for SimplyPrint folder existing prior to trying to remove files in it

+++ octoprint_simplyprint/comm/simplyprint.py
@@ -90,6 +90,7 @@ class SimplyPrintComm:
         self.user_input_required = False
         self.downloading = False
         self.download_status = None
+        self.health_counter = 0
 
         self.last_connection_attempt = time.time()
         self.first = True
@@ -335,8 +336,12 @@ class SimplyPrintComm:
             if self.user_input_required:
                 url += "&userinputrequired"
             url += "&extra=" + url_quote(json.dumps(to_set))
-            resources = Monitor(self._logger)
-            url += "&health=" + url_quote(json.dumps(resources.get_all_resources()))
+            if self.health_counter > 9:
+                self.health_counter = 0
+                resources = Monitor(self._logger)
+                url += "&health=" + url_quote(json.dumps(resources.get_all_resources()))
+            else:
+                self.health_counter = self.health_counter + 1
 
         return self._simply_get(url)
 
@@ -1329,10 +1334,11 @@ class SimplyPrintComm:
         local = FileDestinations.LOCAL
 
         # Delete any old files in SimplyPrint folder
-        files = self.plugin._file_manager.list_files(local, "SimplyPrint")
-        for file, data in files["local"].items():
-            # assume we only upload to this folder and delete everything...
-            self.plugin._file_manager.remove_file(local, data["path"])
+        if self.plugin._file_manager.folder_exists(local, "SimplyPrint"):
+            files = self.plugin._file_manager.list_files(local, "SimplyPrint")
+            for file, data in files["local"].items():
+                # assume we only upload to this folder and delete everything...
+                self.plugin._file_manager.remove_file(local, data["path"])
 
         if new_filename is None:
             new_filename = str(uuid.uuid1())
